refactor(Textarea): eliminar archivo interface y agregar su contenido en index
